requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.1.0
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
testcontainers==4.5.1
typing_extensions==4.12.1
urllib3==2.2.1
validators==0.28.3
wrapt==1.16.0
yarl==1.9.4
